requirements.txt
appdirs==1.4.3
asgiref==3.2.10
CacheControl==0.12.6
cachetools==4.1.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.1
google-auth==1.23.0
google-auth-oauthlib==0.4.2
gspread==3.6.0
html5lib==1.0.1
httplib2==0.18.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
mysqlclient==2.0.1
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
progress==1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==0.10.1
pyparsing==2.4.6
python-decouple==3.3
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
webencodings==0.5.1
